0-langchain-chatbot/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.36",
    "python-dotenv>=1.0",
    "boto3>=1.34",
    "langchain>=0.2",
    "langchain-core>=0.2",
    "langchain-aws>=0.2",
    "langchain-mcp-adapters>=0.0.11",
    "mcp>=1.0.0",
    "awslabs-aws-api-mcp-server>=0.2.13",
    "awslabs-aws-documentation-mcp-server>=1.1.7",
]
